requirements.txt
matplotlib==3.8.0
numpy==1.23.5
PyQt6==6.6.1
pyserial==3.5
scipy==1.12.0
sounddevice==0.4.6
pyqtdarktheme==2.1.0
